requirements.txt
Django==1.8.2
PyYAML==3.11
argparse==1.2.1
awsebcli==3.7.3
backports.ssl-match-hostname==3.5.0.1
blessed==1.9.5
botocore==1.4.0
cement==2.4.0
colorama==0.3.3
coverage==3.7.1
dj-database-url==0.3.0
django-cors-headers==1.1.0
django-hstore==1.3.8
djangorestframework==3.1.1
djangorestframework-hstore==1.2
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.3
pdb==0.1
pep8==1.6.2
psycopg2==2.6
python-dateutil==2.5.0
python-gnupg==0.3.7
pytz==2015.7
requests==2.6.2
six==1.10.0
texttable==0.8.4
wcwidth==0.1.5
websocket-client==0.35.0
wsgiref==0.1.2
